+++ lib/Core/TerriaError.ts
@@ -52,8 +52,8 @@ export interface TerriaErrorOptions {
   /** The object that raised the error. */
   sender?: unknown;
 
-  /** True if error message should be shown to user *regardless* of error severity. If this is undefined, then error severity will be used to determine if shouldRaiseToUser (severity `Error` are presented to the user. `Warning` will just be printed to console) */
-  shouldRaiseToUser?: boolean;
+  /** True if error message should be shown to user *regardless* of error severity. If this is undefined, then error severity will be used to determine if forceRaiseToUser (severity `Error` are presented to the user. `Warning` will just be printed to console) */
+  forceRaiseToUser?: boolean;
 
   /** True if the user has seen this error; otherwise, false. */
   raisedToUser?: boolean;
@@ -109,8 +109,6 @@ export function parseOverrides(
 export default class TerriaError {
   private readonly _message: string | I18nTranslateString;
   private readonly _title: string | I18nTranslateString;
-  /** Override shouldRaiseToUser (see `get shouldRaiseToUser()`) */
-  private _shouldRaiseToUser: boolean | undefined;
   private _raisedToUser: boolean;
 
   readonly importance: number = 0;
@@ -120,6 +118,8 @@ export default class TerriaError {
   readonly originalError?: (TerriaError | Error)[];
   readonly stack: string;
 
+  /** Override shouldRaiseToUser (see `get shouldRaiseToUser()`) */
+  forceRaiseToUser: boolean | undefined;
   @observable showDetails: boolean;
 
   /**
@@ -216,9 +216,9 @@ export default class TerriaError {
         ? TerriaErrorSeverity.Error
         : TerriaErrorSeverity.Warning;
 
-    // shouldRaiseToUser will be true if at least one error includes shouldRaiseToUser = true
-    const shouldRaiseToUser = filteredErrors
-      .map(error => error._shouldRaiseToUser ?? false)
+    // forceRaiseToUser will be true if at least one error includes forceRaiseToUser = true
+    const forceRaiseToUser = filteredErrors
+      .map(error => error.forceRaiseToUser ?? false)
       .includes(true);
 
     return new TerriaError({
@@ -229,7 +229,7 @@ export default class TerriaError {
       // Add original errors and overrides
       originalError: filteredErrors,
       severity,
-      shouldRaiseToUser,
+      forceRaiseToUser,
       ...parseOverrides(overrides)
     });
   }
@@ -239,7 +239,7 @@ export default class TerriaError {
     this._title = options.title ?? { key: "core.terriaError.defaultTitle" };
     this.sender = options.sender;
     this._raisedToUser = options.raisedToUser ?? false;
-    this._shouldRaiseToUser = options.shouldRaiseToUser;
+    this.forceRaiseToUser = options.forceRaiseToUser;
     this.importance = options.importance ?? 0;
     this.showDetails = options.showDetails ?? false;
 
@@ -276,15 +276,11 @@ export default class TerriaError {
     return resolveI18n(this._title);
   }
 
-  set shouldRaiseToUser(s: boolean | undefined) {
-    this._shouldRaiseToUser = s;
-  }
-
-  /** True if `severity` is `Error` and the error hasn't been raised yet - or return this._shouldRaiseToUser if it is defined */
+  /** True if `severity` is `Error` and the error hasn't been raised yet - or return this.forceRaiseToUser if it is defined */
   get shouldRaiseToUser() {
     return (
-      // Return this._shouldRaiseToUser override if it is defined
-      this._shouldRaiseToUser ??
+      // Return this.forceRaiseToUser override if it is defined
+      this.forceRaiseToUser ??
       // Otherwise, we should raise the error if it hasn't already been raised and the severity is ERROR
       (!this.raisedToUser &&
         (typeof this.severity === "function"
@@ -333,7 +329,7 @@ export default class TerriaError {
       sender: this.sender,
       originalError: this,
       severity: this.severity,
-      shouldRaiseToUser: this._shouldRaiseToUser,
+      forceRaiseToUser: this.forceRaiseToUser,
       ...parseOverrides(overrides)
     });
   }