+++ CHANGES.md
@@ -3,6 +3,7 @@ Change Log
 
 #### next release (8.1.5)
 
+* Rename `TerriaError._shouldRaiseToUser` to `forceRaiseToUser`
 * [The next improvement]
 
 #### 8.1.4